feat(tauri): extract process_kill helpers (#1060)
